requirements.txt
et_xmlfile==2.0.0
numpy==2.2.6
openpyxl==3.1.5
pandas==2.2.3
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyodbc==5.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.41
tqdm==4.67.1
typing_extensions==4.14.0
tzdata==2025.2
